# Faster conversion between row and column representations of sparse matrices