A similar change was included in one of the earlier batches of sparse matrix updates but it seemed to have been lost in the shuffle. For a simple benchmark it provides a ~5x speedup.

This change improves performance when converting between row and column representations of sparse matrices. The previous version of getColumns/getRows involved converting a `SparseRowMatrix` to a `SparseColumnMatrix` and then calling getColumns or getRows on the new matrix. However, the conversion between matrices was inefficient where the `unsafeSet` method was being called on the new matrix to transfer values over which includes O(lg n) operations for every call.

I ran a simple benchmark with Criterium:

```
WARNING: Final GC required 1.0204683913295038 % of runtime
Evaluation count : 9120 in 60 samples of 152 calls.
             Execution time mean : 6.880668 ms
    Execution time std-deviation : 192.238086 µs
   Execution time lower quantile : 6.624899 ms ( 2.5%)
   Execution time upper quantile : 7.242259 ms (97.5%)
                   Overhead used : 2.279281 ns

Found 1 outliers in 60 samples (1.6667 %)
	low-severe	 1 (1.6667 %)
 Variance from outliers : 15.7399 % Variance is moderately inflated by outliers

WARNING: Final GC required 1.601230154441423 % of runtime
Evaluation count : 1680 in 60 samples of 28 calls.
             Execution time mean : 35.976748 ms
    Execution time std-deviation : 853.381594 µs
   Execution time lower quantile : 34.495605 ms ( 2.5%)
   Execution time upper quantile : 37.729319 ms (97.5%)
                   Overhead used : 2.279281 ns

Found 1 outliers in 60 samples (1.6667 %)
	low-severe	 1 (1.6667 %)
 Variance from outliers : 11.0367 % Variance is moderately inflated by outliers
```

The code used to run the benchmarks:
```
  (require '[criterium.core :as crit])
  (require '[clojure.core.matrix :as mx])
  (require '[munge.matrix :as mmx])
  (import '[mikera.matrixx.impl SparseColumnMatrix])
  
  (mx/set-current-implementation :vectorz)

  (defn random-cells [num-cells max-index]
    (->> (map vector (repeatedly #(-> (rand) (* max-index) Math/floor int)) (repeatedly rand))
         (take num-cells)))

  
  (def m (mmx/sparse-matrix 1000 1000 (repeatedly 1000 #(random-cells 100 1000))))

  (crit/bench (.toSparseColumnMatrixAlt m)
  (crit/bench (.toSparseColumnMatrix m))
```

That code isn't directly runnable for you but is provided for reference in case you want to verify the performance gain.