@mikera 2015-04-23
Looks good. Smart changes. Thanks!